rest_service/requirements.txt
certifi==2026.4.22
charset-normalizer==3.4.7
coverage==7.13.5
docopt==0.6.2
idna==3.13
iniconfig==2.3.0
mysql-connector-python==9.7.0
packaging==26.2
pluggy==1.6.0
Pygments==2.20.0
pytest==9.0.3
pytest-cov==7.1.0
requests==2.33.1
urllib3==2.6.3
yarg==0.1.10
